Add oxen repo by default; sid workaround; clang-14
Having the oxen repo lets us avoid building oxenmq, etc. which reduces
build time a bit.  It will *also* catch cases where we *need* a version
bump but haven't properly increased the required version (e.g. because
it'll use the release version that doesn't have what we want).

However, lsb_release -sc is currently broken on sid (it returns n/a
instead of sid) which breaks its use here, so this also explicitly
specifies `sid` for the sid builds as a workaround.

This also updates the clang builds to the latest (14).