Fix bootstrap loading to allow an empty override
You should be able to give an explicit file containing `le` to indicate
an empty list of bootstraps.  Currently that doesn't work because the
code goes on to load defaults if the file didn't load anything.

This fixes it to load defaults only when no file is given in the config.