Fix race condition in tun IP assignment
If two session-routers start at the same time they can end up assigning
the same network block to their respective tun devices, which is bad.
This isn't only a race: this can also happen if two session-routers are
configured with a specified overlapping range, in which case neither
notices and at least one will be broken.

This adds two mitigations:
- after assigning IPs to the tun, we loop through network devices to
  make sure there is no overlap; if there is, we throw and abort
  startup.
- for service nodes (where this is most common) we add a small random
  delay (0-25ms) to randomize the startup sequence across difference
  nodes to lower the probability of hitting it.  (And if you do hit it,
  session-router exits, restarts itself, and then should pick a
  non-conflicting range).